fix(exporter/prometheus): use synctest for background cleanup test
Replace time-dependent require.Eventually with testing/synctest for
deterministic fake-clock testing. Moves TestPrometheusExporter_BackgroundCleanup
to a separate file with //go:build goexperiment.synctest build tag, matching
the pattern used in deltatocumulativeprocessor/stale_test.go.
